AXIS2-4200: Merged r896320 to the 1.5 branch.

+++ modules/adb/src/org/apache/axis2/databinding/utils/ConverterUtil.java
@@ -1282,8 +1282,8 @@ public class ConverterUtil {
      * @param value
      * @return 0 if equal , + value if greater than , - value if less than
      */
-    public static int compare(BigInteger binBigInteger, String value) {
-        return binBigInteger.intValue() - Integer.parseInt(value);
+    public static long compare(BigInteger binBigInteger, String value) {
+        return binBigInteger.longValue() - Long.parseLong(value);
     }
 
     /**
